feat(sort): extract framework-light sort CLI into standalone crates (#446)
Extract the `fgumi sort` command and its pipeline steps into three new
crates so sort-only downstreams (e.g. mako) can depend on sort without
compiling the UMI/consensus stack:

- fgumi-cli-common: shared CLI infrastructure (Command trait,
  OperationTimer, memory/compression options, validation, memory
  detection).
- fgumi-pipeline-io: generic BGZF source/sink steps (ReadBgzfBlocks,
  WriteBgzfFile) plus the typed-step sort steps (SortBamFile,
  SortAndSpill, SortMerge, SortSpillDecompress).
- fgumi-sort-cli: the `Sort` command, options, finalize hooks, and step
  factory. `Sort::execute` builds the typed-step pipeline directly via
  `fgumi_pipeline_core::Pipeline::builder` rather than the umbrella's
  monolithic `ChainBuilder`, keeping the dependency graph free of
  fgumi-consensus, fgumi-umi, and fgumi-simd-fastq.

The umbrella `fgumi` crate re-exports the moved items at their original
paths so runall, chains, merge, and main.rs compile unchanged; the
chains layer imports the sort step and factory from the new crates
one-way. Sorted output is byte-identical across coordinate, queryname,
queryname::natural, and template-coordinate orders (and stdin), so this
is a behavior-preserving extraction.
