+++ crates/fgumi-pipeline-io/src/sink/mod.rs
@@ -1,0 +1,1 @@
+pub mod write_bgzf;
