+++ crates/fgumi-pipeline-io/src/sort/and_spill/tests.rs
@@ -1,0 +1,1 @@
+// No unit tests for and_spill directly; integration tests live in sort/tests.rs.
