+++ crates/fgumi-pipeline-io/src/sort/merge/tests.rs
@@ -1,0 +1,1 @@
+// No unit tests for merge directly; integration tests live in sort/tests.rs.
