+++ crates/fgumi-pipeline-io/src/sort/spill_decompress/tests.rs
@@ -1,0 +1,1 @@
+// No unit tests for spill_decompress directly; integration tests live in sort/tests.rs.
