+++ crates/fgumi-pipeline-io/src/source/mod.rs
@@ -1,0 +1,1 @@
+pub mod read_bam;
