+++ src/lib/pipeline/steps/types.rs
@@ -17,244 +17,20 @@
 //! `Vec<Template>`); the new framework's batch carries metadata alongside
 //! the templates.
 
-use fgumi_raw_bam::RawRecord;
-
 use crate::pipeline::core::item::{HeapSize, Ordered};
 use crate::template::Template;
 use fgumi_bam_io::DecodedRecord;
 
 // ─────────────────────────────────────────────────────────────────────────────
-// BgzfBlock — raw compressed BGZF block + read-order serial.
+// Light flowing types (BgzfBlock, DecompressedBlock, RecordBatch,
+// RecordBatchBuilder) moved to the `fgumi-pipeline-io` crate. Re-export them so
+// `crate::pipeline::steps::types::*` paths keep resolving and there is ONE
+// canonical definition of each. The heavy types below (BamTemplateBatch,
+// DecodedRecordBatch, SamChunk) stay in the umbrella because they depend on
+// `crate::template::Template`.
 // ─────────────────────────────────────────────────────────────────────────────
 
-/// Raw compressed BGZF block + parsed metadata. Carries read-order serial.
-///
-/// Sentinel/EOF blocks have `bytes` containing the 28-byte BGZF EOF marker
-/// and `uncompressed_size = 0`.
-#[derive(Debug)]
-pub struct BgzfBlock {
-    /// Read-order serial. Set by `ReadBgzfBlocks` based on block read index.
-    pub batch_serial: u64,
-    pub bytes: Vec<u8>,
-    /// Decompressed size, parsed from the BGZF block header.
-    pub uncompressed_size: u32,
-}
-
-impl HeapSize for BgzfBlock {
-    fn heap_size(&self) -> usize {
-        self.bytes.len()
-    }
-}
-
-impl Ordered for BgzfBlock {
-    fn ordinal(&self) -> u64 {
-        self.batch_serial
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// DecompressedBlock — raw bytes from a BGZF decompression, record-aligned
-// or not.
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// Decompressed bytes from one or more BGZF blocks. Carries a serial for
-/// ordering purposes; record alignment is the consumer's responsibility
-/// (see `FindBamBoundaries`).
-#[derive(Debug)]
-pub struct DecompressedBlock {
-    pub batch_serial: u64,
-    pub bytes: Vec<u8>,
-}
-
-impl HeapSize for DecompressedBlock {
-    fn heap_size(&self) -> usize {
-        self.bytes.len()
-    }
-}
-
-impl Ordered for DecompressedBlock {
-    fn ordinal(&self) -> u64 {
-        self.batch_serial
-    }
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// RecordBatch — parsed BAM records grouped into a batch.
-//
-// Internal representation: a single flat `backing` buffer holding all record
-// bodies concatenated, plus per-record (start, end) byte ranges into it. This
-// avoids the per-record `Vec<u8>` allocation that an
-// earlier `Vec<RawRecord>` shape required during high-throughput parse and
-// merge-emit paths (ParseBamRecords parses N records in a block; Sort's
-// `merge_into` emits N records per batch — both can run zero-alloc per record
-// at the inner loop, with one `Vec` allocation per batch instead).
-//
-// Public API:
-//   - `iter_record_bytes()`           — borrowing iterator of `&[u8]` views
-//   - `len()`, `is_empty()`           — record count
-//   - `total_bytes()`, `batch_serial()` — accessors
-//   - [`RecordBatchBuilder`]          — emit-side builder (`push_record_bytes`)
-//   - `RecordBatch::from_parsed(...)` — ParseBam-side constructor
-//   - `RecordBatch::new(...)`         — convenience for tests; serializes a
-//                                       `Vec<RawRecord>` into the flat repr.
-// ─────────────────────────────────────────────────────────────────────────────
-
-/// A batch of parsed BAM records, stored as a flat backing buffer + per-record
-/// `(start, end)` ranges. See module-level note for rationale.
-#[derive(Debug)]
-pub struct RecordBatch {
-    batch_serial: u64,
-    /// All record bodies concatenated, in batch order.
-    backing: Vec<u8>,
-    /// (start, end) byte ranges into `backing`, one per record. `u32` is
-    /// sufficient since BAM record body size is u32-bounded and we never
-    /// concat more than `u32::MAX` bytes per batch in practice
-    /// (`tuning.per_step_byte_limit` defaults to 4 MiB).
-    ranges: Vec<(u32, u32)>,
-}
-
-impl RecordBatch {
-    /// Construct a batch from a pre-parsed backing buffer and `(start, end)`
-    /// ranges. Used by `ParseBamRecords` to wrap a `DecompressedBlock`'s
-    /// bytes without re-allocating per record.
-    #[must_use]
-    pub fn from_parsed(batch_serial: u64, backing: Vec<u8>, ranges: Vec<(u32, u32)>) -> Self {
-        Self { batch_serial, backing, ranges }
-    }
-
-    /// Convenience constructor: serializes a slice of `RawRecord`s into the
-    /// flat representation. Intended for tests; production emitters should
-    /// use [`RecordBatchBuilder`] to avoid the intermediate `Vec<RawRecord>`
-    /// allocation entirely.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the concatenated record bodies exceed `u32::MAX` bytes
-    /// (the `(start, end)` range type for each record is `(u32, u32)`).
-    /// In practice this is unreachable for typical BAM workloads —
-    /// `tuning.per_step_byte_limit` defaults to 4 MiB.
-    #[must_use]
-    pub fn new(batch_serial: u64, records: &[RawRecord]) -> Self {
-        let total: usize = records.iter().map(RawRecord::len).sum();
-        let mut backing = Vec::with_capacity(total);
-        let mut ranges = Vec::with_capacity(records.len());
-        for rec in records {
-            let start = u32::try_from(backing.len()).expect("backing fits in u32");
-            backing.extend_from_slice(rec.as_ref());
-            let end = u32::try_from(backing.len()).expect("backing fits in u32");
-            ranges.push((start, end));
-        }
-        Self { batch_serial, backing, ranges }
-    }
-
-    /// Self-managed ordinal.
-    #[must_use]
-    pub fn batch_serial(&self) -> u64 {
-        self.batch_serial
-    }
-
-    /// Number of records in the batch.
-    #[must_use]
-    pub fn len(&self) -> usize {
-        self.ranges.len()
-    }
-
-    /// `true` iff the batch contains zero records.
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.ranges.is_empty()
-    }
-
-    /// Total bytes across all record bodies.
-    #[must_use]
-    pub fn total_bytes(&self) -> usize {
-        self.backing.len()
-    }
-
-    /// Iterate the record bodies as borrowed byte slices into the backing
-    /// buffer. Zero-allocation per record.
-    pub fn iter_record_bytes(&self) -> impl Iterator<Item = &[u8]> + '_ {
-        let backing = &self.backing[..];
-        self.ranges.iter().map(move |&(s, e)| &backing[s as usize..e as usize])
-    }
-}
-
-/// Builder for emit-side `RecordBatch` construction. Used by `Sort`'s
-/// `merge_into` closure: append record bodies one at a time into a shared
-/// backing buffer; finalize with [`Self::build`]. One backing allocation
-/// per emitted batch instead of one per record.
-#[derive(Debug)]
-pub struct RecordBatchBuilder {
-    batch_serial: u64,
-    backing: Vec<u8>,
-    ranges: Vec<(u32, u32)>,
-}
-
-impl RecordBatchBuilder {
-    /// Build an empty builder with reserved capacity for `bytes_cap` bytes of
-    /// record body and `records_cap` records.
-    #[must_use]
-    pub fn with_capacity(batch_serial: u64, bytes_cap: usize, records_cap: usize) -> Self {
-        Self {
-            batch_serial,
-            backing: Vec::with_capacity(bytes_cap),
-            ranges: Vec::with_capacity(records_cap),
-        }
-    }
-
-    /// Append one record's body bytes. Memcpys into the backing buffer; no
-    /// per-record allocation.
-    ///
-    /// # Panics
-    ///
-    /// Panics if accumulated bytes would exceed `u32::MAX` (each
-    /// record's `(start, end)` range is `(u32, u32)`). Unreachable in
-    /// practice — `tuning.per_step_byte_limit` defaults to 4 MiB.
-    pub fn push_record_bytes(&mut self, bytes: &[u8]) {
-        let start = u32::try_from(self.backing.len()).expect("backing fits in u32");
-        self.backing.extend_from_slice(bytes);
-        let end = u32::try_from(self.backing.len()).expect("backing fits in u32");
-        self.ranges.push((start, end));
-    }
-
-    /// Number of records appended so far.
-    #[must_use]
-    pub fn len(&self) -> usize {
-        self.ranges.len()
-    }
-
-    /// `true` iff no records have been appended.
-    #[must_use]
-    pub fn is_empty(&self) -> bool {
-        self.ranges.is_empty()
-    }
-
-    /// Total record bytes appended so far.
-    #[must_use]
-    pub fn total_bytes(&self) -> usize {
-        self.backing.len()
-    }
-
-    /// Finalize and produce the `RecordBatch`. Consumes the builder.
-    #[must_use]
-    pub fn build(self) -> RecordBatch {
-        RecordBatch { batch_serial: self.batch_serial, backing: self.backing, ranges: self.ranges }
-    }
-}
-
-impl HeapSize for RecordBatch {
-    fn heap_size(&self) -> usize {
-        // Backing bytes plus the ranges Vec's heap. (Record count × 8 bytes
-        // since each range is `(u32, u32)`.)
-        self.backing.len() + self.ranges.len() * std::mem::size_of::<(u32, u32)>()
-    }
-}
-
-impl Ordered for RecordBatch {
-    fn ordinal(&self) -> u64 {
-        self.batch_serial
-    }
-}
+pub use fgumi_pipeline_io::types::{BgzfBlock, DecompressedBlock, RecordBatch, RecordBatchBuilder};
 
 // ─────────────────────────────────────────────────────────────────────────────
 // DecodedRecordBatch — parsed records with pre-computed `GroupKey`s.
@@ -388,59 +164,9 @@ impl Ordered for SamChunk {
 mod tests {
     use super::*;
 
-    #[test]
-    fn bgzf_block_heap_size_matches_bytes_len() {
-        let b = BgzfBlock { batch_serial: 0, bytes: vec![0u8; 1024], uncompressed_size: 4096 };
-        assert_eq!(b.heap_size(), 1024);
-        assert_eq!(b.ordinal(), 0);
-    }
-
-    #[test]
-    fn decompressed_block_heap_size_matches_bytes_len() {
-        let b = DecompressedBlock { batch_serial: 7, bytes: vec![0u8; 4096] };
-        assert_eq!(b.heap_size(), 4096);
-        assert_eq!(b.ordinal(), 7);
-    }
-
-    #[test]
-    fn record_batch_total_bytes_sums_record_lengths() {
-        // For the size-accounting test we just need records with known
-        // byte lengths; the bytes need not be valid BAM records.
-        let r1: RawRecord = vec![0u8; 100].into();
-        let r2: RawRecord = vec![0u8; 200].into();
-        let batch = RecordBatch::new(3, &[r1, r2]);
-        assert_eq!(batch.len(), 2);
-        assert_eq!(batch.total_bytes(), 300);
-        // `heap_size` includes the ranges Vec (8 bytes per entry).
-        assert_eq!(batch.heap_size(), 300 + 2 * std::mem::size_of::<(u32, u32)>());
-        assert_eq!(batch.ordinal(), 3);
-    }
-
-    #[test]
-    fn record_batch_from_parsed_round_trips_ranges() {
-        let backing = b"AAABBBBCC".to_vec();
-        let ranges = vec![(0u32, 3u32), (3u32, 7u32), (7u32, 9u32)];
-        let batch = RecordBatch::from_parsed(11, backing, ranges);
-        let got: Vec<&[u8]> = batch.iter_record_bytes().collect();
-        assert_eq!(got, vec![&b"AAA"[..], &b"BBBB"[..], &b"CC"[..]]);
-        assert_eq!(batch.len(), 3);
-        assert_eq!(batch.total_bytes(), 9);
-        assert_eq!(batch.batch_serial(), 11);
-    }
-
-    #[test]
-    fn record_batch_builder_collects_records() {
-        let mut b = RecordBatchBuilder::with_capacity(0, 64, 4);
-        b.push_record_bytes(&[0u8; 50]);
-        b.push_record_bytes(&[0u8; 75]);
-        assert_eq!(b.len(), 2);
-        assert!(!b.is_empty());
-        assert_eq!(b.total_bytes(), 125);
-        let batch = b.build();
-        let got: Vec<usize> = batch.iter_record_bytes().map(<[u8]>::len).collect();
-        assert_eq!(got, vec![50, 75]);
-        assert_eq!(batch.heap_size(), 125 + 2 * std::mem::size_of::<(u32, u32)>());
-    }
+    // Tests for the light types (BgzfBlock, DecompressedBlock, RecordBatch,
+    // RecordBatchBuilder) moved with those types to the `fgumi-pipeline-io`
+    // crate. Only the heavy-type tests remain below.
 
     #[test]
     fn template_batch_carries_serial() {
